AG-15406: Separate Flash On Update from Animation docs and features matrix
- Move flash-on-update examples from animation/ to their own flash-on-update/ docs page
- Add Flash On Update as a separate entry in the features matrix (enterprise)
- Add flash-on-update to docs nav and module mappings
- Conditionally disable animation when flashOnUpdate is enabled
- Add flash-on-update theme options to ag-charts-types
- Update ag-charts-versions.json for the new page
