+++ packages/ag-charts-website/src/content/docs-nav/nav.json
@@ -274,6 +274,12 @@
                         },
                         {
                             "type": "item",
+                            "title": "Flash On Update",
+                            "path": "flash-on-update",
+                            "isEnterprise": true
+                        },
+                        {
+                            "type": "item",
                             "title": "Navigator",
                             "path": "navigator",
                             "isEnterprise": true
