+++ packages/ag-charts-website/src/content/docs/flash-on-update/_examples/flash-on-update-category/index.html
