+++ packages/ag-charts-website/src/content/docs/flash-on-update/_examples/flash-on-update/main.ts
@@ -48,7 +48,6 @@ const options: AgCartesianChartOptions = {
             },
         },
     },
-    animation: { enabled: false },
     flashOnUpdate: {
         enabled: true,
     },
